refactor: extract player update logic into separate function
